feat: update blog-post path
