+++ refine-react/plugins/inferencer-headless/src/pages/blog-posts/list.tsx
