+++ refine-react/plugins/inferencer/src/pages/blog-posts/create.tsx
