+++ refine-react/plugins/inferencer/src/pages/blog-posts/edit.tsx
