+++ refine-react/plugins/inferencer/src/pages/blog-posts/index.ts
