+++ refine-react/plugins/inferencer/src/pages/blog-posts/list.tsx
