+++ refine-react/plugins/inferencer/src/pages/blog-posts/show.tsx
