+++ refine-react/template/src/App.tsx
@@ -48,10 +48,10 @@ function App() {
                         <%_ if (answers["data-provider"] === 'data-provider-strapi-v4') { _%>
                         {
                             name: "blog-posts",
-                            list: "/blog_posts",
-                            create: "/blog_posts/create",
-                            edit: "/blog_posts/edit/:id",
-                            show: "/blog_posts/show/:id",
+                            list: "/blog-posts",
+                            create: "/blog-posts/create",
+                            edit: "/blog-posts/edit/:id",
+                            show: "/blog-posts/show/:id",
                             meta: {
                                 canDelete: true,
                             },
@@ -59,10 +59,10 @@ function App() {
                         <%_ } else { _%>
                         {
                             name: "blog_posts",
-                            list: "/blog_posts",
-                            create: "/blog_posts/create",
-                            edit: "/blog_posts/edit/:id",
-                            show: "/blog_posts/show/:id",
+                            list: "/blog-posts",
+                            create: "/blog-posts/create",
+                            edit: "/blog-posts/edit/:id",
+                            show: "/blog-posts/show/:id",
                             meta: {
                                 canDelete: true,
                             },
@@ -109,7 +109,7 @@ function App() {
                                 <NavigateToResource resource="blog_posts" />
                             <%_ } _%>
                         } />
-                        <Route path="/blog_posts">
+                        <Route path="/blog-posts">
                             <Route index element={<BlogPostList />} />
                             <Route path="create" element={<BlogPostCreate />} />
                             <Route path="edit/:id" element={<BlogPostEdit />} />
@@ -163,7 +163,7 @@ function App() {
                                 <NavigateToResource resource="blog_posts" />
                             <%_ } _%>
                         } />
-                        <Route path="/blog_posts">
+                        <Route path="/blog-posts">
                             <Route index element={<BlogPostList />} />
                             <Route path="create" element={<BlogPostCreate />} />
                             <Route path="edit/:id" element={<BlogPostEdit />} />
