+++ refine-remix/plugins/inferencer-headless/app/routes/_layout.blog-posts.create.tsx
